style: properly align rows in viewer
We do not use the Wide grid mode (https://carbondesignsystem.com/guidelines/2x-grid/implementation#the-wide-grid), so a slight negative margin is added to both side. We want to prevent that when embedding the viewer inside applications.

We still apply the margin to the playground to not waste space. 

Related to https://github.com/camunda/tasklist/pull/2746