feat(viewer): support `properties.disabled`
BREAKING CHANGES:

We changed the behavior when providing the `readOnly` property to a Form. From this version, the form fields will be rendered as `readOnly` if the property is set. Previously, the form fields were rendered as `disabled`. To restore the same behavior, please use the `disabled` property instead.

```js
const form = new Form({
  container: document.querySelector('#form'),
  properties: {
    disabled: true
  }
});
```